+++ datastore_server/source/java/ch/ethz/sis/openbis/generic/server/dss/plugins/DataSetRegistrationCleanUpTask.java
@@ -18,16 +18,27 @@ package ch.ethz.sis.openbis.generic.server.dss.plugins;
 
 import java.io.File;
 import java.io.FileFilter;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.LinkOption;
+import java.nio.file.Path;
+import java.nio.file.attribute.FileTime;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
+import java.util.Collection;
 import java.util.Date;
+import java.util.List;
 import java.util.Properties;
 import java.util.Set;
 import java.util.TreeSet;
+import java.util.function.Consumer;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import org.apache.commons.lang3.time.DateUtils;
 import org.apache.log4j.Logger;
 
+import ch.systemsx.cisd.base.exceptions.CheckedExceptionTunnel;
 import ch.systemsx.cisd.common.filesystem.FileUtilities;
 import ch.systemsx.cisd.common.logging.Log4jSimpleLogger;
 import ch.systemsx.cisd.common.logging.LogCategory;
@@ -39,13 +50,20 @@ import ch.systemsx.cisd.common.utilities.SystemTimeProvider;
 import ch.systemsx.cisd.etlserver.ETLDaemon;
 import ch.systemsx.cisd.etlserver.ThreadParameters;
 import ch.systemsx.cisd.etlserver.TopLevelDataSetRegistratorGlobalState;
+import ch.systemsx.cisd.etlserver.registrator.v2.DataSetRegistrationService;
+import ch.systemsx.cisd.openbis.dss.generic.server.plugins.standard.IngestionService;
+import ch.systemsx.cisd.openbis.dss.generic.shared.IEncapsulatedOpenBISService;
 import ch.systemsx.cisd.openbis.dss.generic.shared.ServiceProvider;
+import ch.systemsx.cisd.openbis.dss.generic.shared.utils.SegmentedStoreUtils;
+import ch.systemsx.cisd.openbis.generic.shared.basic.dto.AbstractExternalData;
+import ch.systemsx.cisd.openbis.generic.shared.basic.dto.DatabaseInstance;
 
 /**
  * @author Franz-Josef Elmer
  */
 public class DataSetRegistrationCleanUpTask implements IMaintenanceTask
 {
+    public static final String ELN_TEMP_DIR_FOR_REGISTRATION = "tmp_eln";
 
     public static final String DEFAULT_MAINTENANCE_TASK_NAME = "data-set-registration-clean-up-task";
 
@@ -54,6 +72,8 @@ public class DataSetRegistrationCleanUpTask implements IMaintenanceTask
     public static final String MINIMUM_AGE_IN_DAYS = "minimum-age-in-days";
 
     public static final int DEFAULT_MINIMUM_AGE_IN_DAYS = 30;
+
+    private static final int SHARDING_LEVELS = 3;
 
     private static final Logger operationLog = LogFactory.getLogger(LogCategory.OPERATION,
             DataSetRegistrationCleanUpTask.class);
@@ -66,33 +86,44 @@ public class DataSetRegistrationCleanUpTask implements IMaintenanceTask
 
     private ITimeProvider timeProvider;
 
+    private IEncapsulatedOpenBISService openBisService;
+
     private MonitoredDirectories preStagingDirectories;
 
     private MonitoredDirectories stagingDirectories;
 
     private MonitoredDirectories preCommitDirectories;
 
+    private MonitoredDirectories elnTmpDirectories;
     private Log4jSimpleLogger simpleLogger;
+
+    private SimpleDateFormat dateFormat;
+
+    private int minimumAgeInDays;
+
 
     public DataSetRegistrationCleanUpTask()
     {
     }
 
     DataSetRegistrationCleanUpTask(ThreadParameters[] threadParameters, File storeRoot,
-            ITimeProvider timeProvider)
+            ITimeProvider timeProvider, IEncapsulatedOpenBISService openBisService)
     {
         this.threadParameters = threadParameters;
         this.storeRoot = storeRoot;
         this.timeProvider = timeProvider;
+        this.openBisService = openBisService;
     }
 
     @Override
     public void setUp(String pluginName, Properties properties)
     {
-        minimumAge = PropertyUtils.getInt(properties, MINIMUM_AGE_IN_DAYS, DEFAULT_MINIMUM_AGE_IN_DAYS) * DateUtils.MILLIS_PER_DAY;
+        minimumAgeInDays = PropertyUtils.getInt(properties, MINIMUM_AGE_IN_DAYS, DEFAULT_MINIMUM_AGE_IN_DAYS);
+        minimumAge = minimumAgeInDays * DateUtils.MILLIS_PER_DAY;
         preStagingDirectories = new MonitoredDirectories("pre-staging");
         stagingDirectories = new MonitoredDirectories("staging");
         preCommitDirectories = new MonitoredDirectories("pre-commit");
+        elnTmpDirectories = new MonitoredDirectories("ELN temporary registration");
         File root = getStoreRoot();
         for (ThreadParameters parameters : getThreadParameters())
         {
@@ -101,8 +132,12 @@ public class DataSetRegistrationCleanUpTask implements IMaintenanceTask
             preStagingDirectories.add(TopLevelDataSetRegistratorGlobalState.getPreStagingDir(root, shareId, tp));
             stagingDirectories.add(TopLevelDataSetRegistratorGlobalState.getStagingDir(root, shareId, tp));
             preCommitDirectories.add(TopLevelDataSetRegistratorGlobalState.getPreCommitDir(root, shareId, tp));
+            elnTmpDirectories.add(new File(root, shareId + "/" + IngestionService.AGGREGATION_SERVICE_SCRATCH_DIR_NAME
+                    + "/" + IngestionService.INCOMING_DIR + "/" + ELN_TEMP_DIR_FOR_REGISTRATION));
         }
         simpleLogger = new Log4jSimpleLogger(operationLog);
+        dateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+
     }
 
     @Override
@@ -110,28 +145,112 @@ public class DataSetRegistrationCleanUpTask implements IMaintenanceTask
     {
         long now = getTimeProvider().getTimeInMilliseconds();
         deleteOldFiles(preStagingDirectories, new PreStagingFileFilter(now, minimumAge));
-        deleteOldFiles(stagingDirectories, new AgeBasedFileFilter(now, minimumAge));
-        deleteOldFiles(preCommitDirectories, new AgeBasedFileFilter(now, minimumAge));
+        FileFilter ageBasedFileFilter = new AgeBasedFileFilter(now, minimumAge);
+        deleteOldFiles(stagingDirectories, ageBasedFileFilter);
+        deleteOldFiles(preCommitDirectories, ageBasedFileFilter);
+        deleteOldFiles(elnTmpDirectories, ageBasedFileFilter);
+        deleteEmptyFoldersInStore(ageBasedFileFilter);
     }
 
     private void deleteOldFiles(MonitoredDirectories dirs, FileFilter filter)
     {
-        SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
         for (File dir : dirs.getDirectories())
         {
             File[] files = dir.listFiles(filter);
             if (files != null && files.length > 0)
             {
-                operationLog.info(dirs.getName() + " directory " + dir.getAbsolutePath()
-                        + " has " + files.length + " files");
+                operationLog.info(dirs.getName() + " directory " + dir.getAbsolutePath() + " has "
+                        + files.length + " files which are older than " + minimumAgeInDays + " days:");
                 for (File file : files)
                 {
                     operationLog.info(file.getName() + " (last modified: "
-                            + format.format(new Date(file.lastModified())) + ")");
+                            + dateFormat.format(new Date(file.lastModified())) + ")");
                     FileUtilities.deleteRecursively(file, simpleLogger);
                     operationLog.info("Stale folder deleted: " + file.getAbsolutePath());
                 }
             }
+        }
+    }
+
+    private void deleteEmptyFoldersInStore(FileFilter fileFilter)
+    {
+        DatabaseInstance homeDatabaseInstance = getOpenBisService().getHomeDatabaseInstance();
+        String instanceHome = homeDatabaseInstance.getUuid();
+        File[] shares = SegmentedStoreUtils.getShares(getStoreRoot());
+        for (File share : shares)
+        {
+            Path shareCore = new File(share, instanceHome).toPath();
+            if (Files.isDirectory(shareCore) == false)
+            {
+                continue;
+            }
+            int shareCorePathLevels = shareCore.getNameCount();
+            EmptyFoldersGatherer gatherer = new EmptyFoldersGatherer(shareCorePathLevels, fileFilter);
+            try (Stream<Path> paths = Files.walk(shareCore, SHARDING_LEVELS + 2))
+            {
+                paths.forEach(gatherer);
+            } catch (Exception e)
+            {
+                throw CheckedExceptionTunnel.wrapIfNecessary(e);
+            }
+            Set<Path> emptyFolders = gatherer.getEmptyFolders();
+            Set<Path> emptyDataSets = gatherer.getEmptyDataSets();
+            emptyFolders.removeAll(emptyDataSets);
+            handleEmptyDataSets(share.getName(), emptyDataSets);
+            handleEmptyFolder(share.getName(), emptyFolders);
+        }
+    }
+
+    private void handleEmptyDataSets(String shareId, Collection<Path> emptyDataSets)
+    {
+        if (emptyDataSets.isEmpty())
+        {
+            return;
+        }
+        operationLog.info(emptyDataSets.size() + " empty data set folders found in share " + shareId
+                + " which are older than " + minimumAgeInDays + " days: " + emptyDataSets);
+        List<String> dataSetCodes = emptyDataSets.stream().map(path -> path.getFileName().toString())
+                .collect(Collectors.toList());
+        List<AbstractExternalData> dataSets = getOpenBisService().listDataSetsByCode(dataSetCodes);
+        Set<String> knownDataSets = collectAsTreeSet(dataSets.stream().map(AbstractExternalData::getCode));
+        for (Path path : emptyDataSets)
+        {
+            if (knownDataSets.contains(path.getFileName().toString()) == false)
+            {
+                deleteFolder(path, "Stale data set folder");
+            }
+        }
+    }
+
+    private void handleEmptyFolder(String shareId, Collection<Path> emptyFolders)
+    {
+        if (emptyFolders.isEmpty())
+        {
+            return;
+        }
+        operationLog.info(emptyFolders.size() + " empty sharding folders found in share " + shareId
+                + " which are older than " + minimumAgeInDays + " days: " + emptyFolders);
+        for (Path path : emptyFolders)
+        {
+            deleteFolder(path, "Empty sharding folder");
+        }
+    }
+
+    private void deleteFolder(Path path, String folderType)
+    {
+        String info = createLastModifiedInfo(path);
+        FileUtilities.deleteRecursively(path.toFile(), simpleLogger);
+        operationLog.info(folderType + " deleted: " + path.toAbsolutePath() + info);
+    }
+
+    private String createLastModifiedInfo(Path path)
+    {
+        try
+        {
+            return " (last modified: " + Files.getLastModifiedTime(path) + ")";
+        } catch (IOException e)
+        {
+            return "";
         }
     }
 
@@ -160,6 +279,76 @@ public class DataSetRegistrationCleanUpTask implements IMaintenanceTask
             timeProvider = SystemTimeProvider.SYSTEM_TIME_PROVIDER;
         }
         return timeProvider;
+    }
+
+    private IEncapsulatedOpenBISService getOpenBisService()
+    {
+        if (openBisService == null)
+        {
+            openBisService = ServiceProvider.getOpenBISService();
+        }
+        return openBisService;
+    }
+
+    private final class EmptyFoldersGatherer implements Consumer<Path>
+    {
+        private final int shareCorePathLevels;
+
+        private final FileFilter fileFilter;
+
+        private final Set<Path> emptyFolders = new TreeSet<>();
+
+        private final Set<Path> emptyDataSets = new TreeSet<>();
+
+        private EmptyFoldersGatherer(int shareCorePathLevels, FileFilter fileFilter)
+        {
+            this.shareCorePathLevels = shareCorePathLevels;
+            this.fileFilter = fileFilter;
+        }
+
+        @Override
+        public void accept(Path path)
+        {
+            if (Files.isDirectory(path, LinkOption.NOFOLLOW_LINKS))
+            {
+                int nameCount = path.getNameCount();
+                if (nameCount == shareCorePathLevels + SHARDING_LEVELS + 1)
+                {
+                    emptyDataSets.add(path);
+                }
+                if (nameCount > shareCorePathLevels + SHARDING_LEVELS + 1)
+                {
+                    emptyDataSets.remove(path.getParent());
+                }
+                if (nameCount > shareCorePathLevels
+                        && nameCount <= shareCorePathLevels + SHARDING_LEVELS + 1)
+                {
+                    emptyFolders.add(path);
+                }
+                emptyFolders.remove(path.getParent());
+            }
+        }
+
+        Set<Path> getEmptyDataSets()
+        {
+            return getOldFolders(emptyDataSets);
+        }
+
+        Set<Path> getEmptyFolders()
+        {
+            return getOldFolders(emptyFolders);
+        }
+
+        private TreeSet<Path> getOldFolders(Set<Path> folders)
+        {
+            return collectAsTreeSet(folders.stream().filter(path -> fileFilter.accept(path.toFile())));
+        }
+
+    }
+
+    private static <T> TreeSet<T> collectAsTreeSet(Stream<T> stream)
+    {
+        return stream.collect(Collectors.toCollection(TreeSet::new));
     }
 
     private static final class MonitoredDirectories