+++ js-test/servers/common/core-plugins/tests/1/as/webapps/openbis-test/html/openbis-test.js
@@ -1481,7 +1481,7 @@ test("createReportFromDataSets()", function() {
 test("listAggregationServices()", function() {
 	createFacadeAndLogin(function(facade) {
 		facade.listAggregationServices(function(response) {
-			assertObjectsCount(response.result, 18);
+			assertObjectsCount(response.result, 20);
 			facade.close();
 		});
 	});